refactor(ui): split main window helpers
Move main window UI wiring, location actions, commands, refresh flow, and display/alert handling into focused mixins while preserving MainWindow as the public class.

Tested: main window regression suite 75 passed, 8 skipped; uv run ruff check src tests scripts; uv run pyright

Confidence: high
