+++ src/accessiweather/ui/main_window.py
@@ -8,24 +8,23 @@ widgets for optimal screen reader compatibility.
 from __future__ import annotations
 
 import logging
-from datetime import datetime
 from typing import TYPE_CHECKING
 
-import wx
-from wx.lib.sized_controls import SizedFrame, SizedPanel
+import wx  # noqa: F401 - public monkeypatch surface
+from wx.lib.sized_controls import SizedFrame
 
-from ..display.presentation.formatters import get_temperature_precision
-from ..runtime_env import is_compiled_runtime
+from ..runtime_env import is_compiled_runtime  # noqa: F401 - public monkeypatch surface
 from ..screen_reader import ScreenReaderAnnouncer
-from ..units import resolve_temperature_unit_preference
-from ..user_manual import open_user_manual
-from ..utils.temperature_utils import format_temperature
-from . import main_window_notification_events
-from .dialogs.location_dialog import show_edit_location_dialog
+from ..user_manual import open_user_manual  # noqa: F401 - public monkeypatch surface
+from .dialogs.location_dialog import show_edit_location_dialog  # noqa: F401
+from .main_window_commands import MainWindowCommandMixin
+from .main_window_display import MainWindowDisplayMixin
+from .main_window_locations import MainWindowLocationMixin
+from .main_window_refresh import MainWindowRefreshMixin
+from .main_window_ui import MainWindowUIMixin
 
 if TYPE_CHECKING:
     from ..app import AccessiWeatherApp
-    from ..models.location import Location
 
 logger = logging.getLogger(__name__)
 
@@ -42,6 +41,9 @@ QUICK_ACTION_LABELS = {
 # Sentinel string used for the "All Locations" summary entry in the dropdown.
 ALL_LOCATIONS_SENTINEL = "All Locations"
 
+# The Help menu label "User &Manual" is created in MainWindowUIMixin and binds
+# to _on_open_user_manual in MainWindowCommandMixin.
+
 
 class _StaleWarningProxy:
     """Delegates SetLabel() calls to the second field of the wx.StatusBar."""
@@ -53,7 +55,14 @@ class _StaleWarningProxy:
         self._window.GetStatusBar().SetStatusText(text, 1)
 
 
-class MainWindow(SizedFrame):
+class MainWindow(
+    MainWindowUIMixin,
+    MainWindowLocationMixin,
+    MainWindowCommandMixin,
+    MainWindowRefreshMixin,
+    MainWindowDisplayMixin,
+    SizedFrame,
+):
     """
     Main application window using plain wxPython.
 
@@ -107,1902 +116,3 @@ class MainWindow(SizedFrame):
 
         # Populate initial data
         self._populate_locations()
-
-    def _create_widgets(self) -> None:
-        """Create all UI widgets."""
-        panel = self.GetContentsPane()
-        panel.SetSizerType("vertical")
-
-        # Location section
-        location_panel = SizedPanel(panel)
-        location_panel.SetSizerType("horizontal")
-        location_panel.SetSizerProps(expand=True)
-
-        wx.StaticText(location_panel, label="Location:")
-        self.location_dropdown = wx.Choice(
-            location_panel,
-            name="Location selection",
-        )
-        self.location_dropdown.SetSizerProps(expand=True, proportion=1)
-
-        # Optional: place Add/Edit/Remove on the location row instead of the
-        # bottom button panel.  Grouped spatially with the dropdown they act on,
-        # at the cost of inserting buttons between the dropdown and the forecast
-        # content in the tab order.
-        location_buttons_on_top = getattr(
-            self.app.config_manager.get_settings(),
-            "location_buttons_on_top",
-            False,
-        )
-        if location_buttons_on_top:
-            self.add_button = wx.Button(location_panel, label=QUICK_ACTION_LABELS["add"])
-            self.edit_button = wx.Button(location_panel, label=QUICK_ACTION_LABELS["edit"])
-            self.remove_button = wx.Button(location_panel, label=QUICK_ACTION_LABELS["remove"])
-
-        # Current conditions section
-        wx.StaticText(panel, label="Current Conditions:")
-        self.current_conditions = wx.TextCtrl(
-            panel,
-            style=wx.TE_MULTILINE | wx.TE_READONLY | wx.TE_RICH2,
-            name="Current weather conditions",
-        )
-        self.current_conditions.SetSizerProps(expand=True, proportion=1)
-
-        # Forecast section
-        self._hourly_forecast_label = wx.StaticText(panel, label="Hourly Forecast:")
-        self.hourly_forecast_display = wx.TextCtrl(
-            panel,
-            style=wx.TE_MULTILINE | wx.TE_READONLY | wx.TE_RICH2,
-            name="Hourly weather forecast",
-        )
-        self.hourly_forecast_display.SetSizerProps(expand=True, proportion=1)
-
-        self._daily_forecast_label = wx.StaticText(panel, label="Daily Forecast:")
-        self.daily_forecast_display = wx.TextCtrl(
-            panel,
-            style=wx.TE_MULTILINE | wx.TE_READONLY | wx.TE_RICH2,
-            name="Daily weather forecast",
-        )
-        self.daily_forecast_display.SetSizerProps(expand=True, proportion=1)
-        self.forecast_display = self.daily_forecast_display
-
-        # Weather alerts section
-        alerts_panel = SizedPanel(panel)
-        alerts_panel.SetSizerType("vertical")
-        alerts_panel.SetSizerProps(expand=True)
-
-        wx.StaticText(alerts_panel, label="Weather Alerts:")
-        self.alerts_list = wx.ListBox(
-            alerts_panel,
-            name="Weather alerts list",
-        )
-        self.alerts_list.SetSizerProps(expand=True, proportion=1)
-
-        self.view_alert_button = wx.Button(alerts_panel, label="View Alert Details")
-        self.view_alert_button.Disable()  # Disabled until alerts are available
-
-        # Event Center section — displayed as "Recent Events" so users
-        # immediately understand it's a reviewable log of recent notifications
-        # and app events, not a live feed.
-        self._event_center_visible = True
-        self._event_center_label = wx.StaticText(panel, label="Recent Events:")
-        self._event_center_tooltip_text = (
-            "Reviewable log of recent weather notifications, AFD updates, "
-            "forecast briefings, and other in-app events."
-        )
-        self._event_center_label.SetToolTip(self._event_center_tooltip_text)
-        self.event_center_display = wx.TextCtrl(
-            panel,
-            style=wx.TE_MULTILINE | wx.TE_READONLY | wx.TE_RICH2,
-            name="Recent events",
-        )
-        self.event_center_display.SetToolTip(self._event_center_tooltip_text)
-        self.event_center_display.SetSizerProps(expand=True, proportion=1)
-
-        # Control buttons
-        button_panel = SizedPanel(panel)
-        button_panel.SetSizerType("horizontal")
-        button_panel.SetSizerProps(expand=True)
-
-        if not location_buttons_on_top:
-            self.add_button = wx.Button(button_panel, label=QUICK_ACTION_LABELS["add"])
-            self.edit_button = wx.Button(button_panel, label=QUICK_ACTION_LABELS["edit"])
-            self.remove_button = wx.Button(button_panel, label=QUICK_ACTION_LABELS["remove"])
-        self.refresh_button = wx.Button(button_panel, label=QUICK_ACTION_LABELS["refresh"])
-        self.explain_button = wx.Button(button_panel, label=QUICK_ACTION_LABELS["explain"])
-        self.discussion_button = wx.Button(button_panel, label=QUICK_ACTION_LABELS["discussion"])
-        self.settings_button = wx.Button(button_panel, label=QUICK_ACTION_LABELS["settings"])
-
-        # Adjacent StaticText explaining why the Forecast Products button is
-        # disabled for non-US locations. Screen readers announce adjacent
-        # StaticText, which is the accessibility affordance for this reason
-        # label (SetName/tooltips are ignored in this project).
-        self.forecast_products_us_only_label = wx.StaticText(
-            panel, label="Forecaster Notes are US-only"
-        )
-        self.forecast_products_us_only_label.Hide()
-
-        # Status bar — two fields: [0] main status, [1] stale/cached warning
-        self.CreateStatusBar(2)
-        self.GetStatusBar().SetStatusWidths([-2, -1])
-        self.stale_warning_label = _StaleWarningProxy(self)
-
-    def _bind_events(self) -> None:
-        """Bind all event handlers."""
-        # Window events
-        self.Bind(wx.EVT_CLOSE, self._on_close)
-        self.Bind(wx.EVT_ICONIZE, self._on_iconize)
-        self.Bind(wx.EVT_SHOW, self._on_window_shown)
-
-        # Location dropdown
-        self.location_dropdown.Bind(wx.EVT_CHOICE, self._on_location_changed)
-
-        # Buttons
-        self.add_button.Bind(wx.EVT_BUTTON, lambda e: self.on_add_location())
-        self.edit_button.Bind(wx.EVT_BUTTON, lambda e: self.on_edit_location())
-        self.remove_button.Bind(wx.EVT_BUTTON, lambda e: self.on_remove_location())
-        self.refresh_button.Bind(wx.EVT_BUTTON, lambda e: self.on_refresh())
-        self.explain_button.Bind(wx.EVT_BUTTON, lambda e: self._on_explain_weather())
-        self.discussion_button.Bind(wx.EVT_BUTTON, lambda e: self._on_discussion())
-        self.settings_button.Bind(wx.EVT_BUTTON, lambda e: self.on_settings())
-        self.view_alert_button.Bind(wx.EVT_BUTTON, self._on_view_alert)
-
-        # Alerts list: open details on double-click or Enter/Space key.
-        # EVT_CHAR_HOOK fires before the screen reader can consume the event,
-        # so Enter/Space reach the handler even when NVDA is in forms mode.
-        self.alerts_list.Bind(wx.EVT_LISTBOX_DCLICK, self._on_view_alert)
-        self.alerts_list.Bind(wx.EVT_CHAR_HOOK, self._on_alerts_list_key)
-
-    def _on_alerts_list_key(self, event) -> None:
-        """Handle key presses in alerts list — Enter/Space opens details."""
-        if event.GetKeyCode() in (wx.WXK_RETURN, wx.WXK_NUMPAD_ENTER, wx.WXK_SPACE):
-            self._on_view_alert(event)
-        else:
-            event.Skip()
-
-    def _on_window_shown(self, event) -> None:
-        """Handle window shown event to set initial focus."""
-        event.Skip()  # Allow event to propagate
-        if event.IsShown():
-            # Use CallLater to ensure focus is set after window is fully ready
-            # A small delay helps screen readers properly announce the focused control
-            wx.CallLater(100, self._set_initial_focus)
-
-    def _set_initial_focus(self) -> None:
-        """Set initial focus to the location dropdown for keyboard accessibility."""
-        try:
-            self.location_dropdown.SetFocus()
-            logger.debug("Initial focus set to location dropdown")
-        except Exception as e:
-            logger.debug(f"Could not set initial focus: {e}")
-
-    def _populate_locations(self) -> None:
-        """
-        Populate the location dropdown with saved locations.
-
-        "All Locations" is always the first entry.  Nationwide is not shown
-        while the All Locations view is active (it is excluded from the
-        per-location summary and does not make sense as a summary entry).
-        """
-        try:
-            locations = self.app.config_manager.get_all_locations()
-            # Exclude Nationwide from the per-location summary list but keep it
-            # in the dropdown as its own selectable entry.
-            location_names = [loc.name for loc in locations]
-            all_names = [ALL_LOCATIONS_SENTINEL] + location_names
-            self.location_dropdown.Clear()
-            self.location_dropdown.Append(all_names)
-
-            # Select current location if set; otherwise land on "All Locations".
-            current = self.app.config_manager.get_current_location()
-            if current and current.name in all_names:
-                idx = all_names.index(current.name)
-                self.location_dropdown.SetSelection(idx)
-                self._update_title_for_location(current.name)
-            else:
-                self.location_dropdown.SetSelection(0)
-                self._update_title_for_location(ALL_LOCATIONS_SENTINEL)
-            self._update_forecast_products_button_state()
-        except Exception as e:
-            logger.error(f"Failed to populate locations: {e}")
-
-    def _update_title_for_location(self, location_name: str | None) -> None:
-        """
-        Set the window title to include the active location.
-
-        The title is used by the taskbar, screen readers on window focus,
-        and Alt+Tab, so surfacing the active location there gives users
-        orientation without needing to read the location dropdown.
-        """
-        base = "AccessiWeather"
-        if location_name and location_name != ALL_LOCATIONS_SENTINEL:
-            self.SetTitle(f"{base} \u2014 {location_name}")
-        else:
-            self.SetTitle(base)
-
-    def _create_menu_bar(self) -> None:
-        """Create the application menu bar."""
-        menu_bar = wx.MenuBar()
-
-        # File menu
-        file_menu = wx.Menu()
-        self._settings_id = wx.ID_PREFERENCES
-        settings_item = file_menu.Append(self._settings_id, "&Settings\tCtrl+S", "Open settings")
-        file_menu.AppendSeparator()
-        self._exit_id = wx.ID_EXIT
-        exit_item = file_menu.Append(self._exit_id, "E&xit\tCtrl+Q", "Exit the application")
-        menu_bar.Append(file_menu, "&File")
-
-        # Location menu
-        location_menu = wx.Menu()
-        self._add_location_id = wx.NewIdRef()
-        add_item = location_menu.Append(
-            self._add_location_id, "&Add Location\tCtrl+L", "Add a new location"
-        )
-        self._edit_location_id = wx.NewIdRef()
-        edit_item = location_menu.Append(
-            self._edit_location_id,
-            "&Edit Location...",
-            "Edit the selected location (e.g. enable Marine Mode)",
-        )
-        self._remove_location_id = wx.NewIdRef()
-        remove_item = location_menu.Append(
-            self._remove_location_id, "&Remove Location\tCtrl+D", "Remove selected location"
-        )
-        menu_bar.Append(location_menu, "&Location")
-
-        # View menu
-        view_menu = wx.Menu()
-        refresh_item = view_menu.Append(wx.ID_REFRESH, "Re&fresh\tF5", "Refresh weather data")
-        view_menu.AppendSeparator()
-        self._explain_id = wx.NewIdRef()
-        explain_item = view_menu.Append(
-            self._explain_id, "&Explain Weather\tCtrl+E", "Get AI explanation of weather"
-        )
-        view_menu.AppendSeparator()
-        self._history_id = wx.NewIdRef()
-        history_item = view_menu.Append(
-            self._history_id, "Weather &History\tCtrl+H", "View weather history"
-        )
-        self._precipitation_timeline_id = wx.NewIdRef()
-        self._precipitation_timeline_item = view_menu.Append(
-            self._precipitation_timeline_id,
-            "Precipitation &Timeline...",
-            "View Pirate Weather minute-by-minute precipitation guidance",
-        )
-        self._toggle_event_center_id = wx.NewIdRef()
-        toggle_event_center_item = view_menu.AppendCheckItem(
-            self._toggle_event_center_id,
-            "Event &Center",
-            "Show or hide the Event Center",
-        )
-        toggle_event_center_item.Check(True)
-        discussion_item = view_menu.Append(
-            wx.ID_ANY,
-            "Forecaster &Notes...",
-            "View NWS forecaster notes (AFD, HWO, SPS)",
-        )
-        aviation_item = view_menu.Append(wx.ID_ANY, "&Aviation Weather...", "View aviation weather")
-        air_quality_item = view_menu.Append(
-            wx.ID_ANY, "Air &Quality...", "View air quality information"
-        )
-        uv_index_item = view_menu.Append(wx.ID_ANY, "&UV Index...", "View UV index information")
-        self._noaa_radio_id = wx.NewIdRef()
-        view_menu.Append(
-            self._noaa_radio_id,
-            "NOAA Weather &Radio...\tCtrl+Shift+R",
-            "Listen to NOAA Weather Radio",
-        )
-        view_menu.AppendSeparator()
-        self._weather_chat_id = wx.NewIdRef()
-        weather_chat_item = view_menu.Append(
-            self._weather_chat_id, "Weather Assistan&t...\tCtrl+T", "Chat with AI weather assistant"
-        )
-        menu_bar.Append(view_menu, "&View")
-
-        # Tools menu
-        tools_menu = wx.Menu()
-        soundpack_item = tools_menu.Append(wx.ID_ANY, "&Soundpack Manager...", "Manage sound packs")
-        menu_bar.Append(tools_menu, "&Tools")
-
-        # Help menu
-        help_menu = wx.Menu()
-
-        # Check for updates - show current channel in label
-        channel = self._get_update_channel()
-        self._check_updates_item = help_menu.Append(
-            wx.ID_ANY,
-            f"Check for &Updates ({channel.title()})...",
-            "Check for application updates",
-        )
-        user_manual_item = help_menu.Append(
-            wx.ID_ANY,
-            "User &Manual",
-            "Open the AccessiWeather user manual",
-        )
-        self._debug_menu_items: dict[str, wx.MenuItem] = {}
-        if getattr(self.app, "debug_mode", False):
-            debug_menu = wx.Menu()
-            self._debug_menu_items["discussion"] = debug_menu.Append(
-                wx.ID_ANY,
-                "Test: &Discussion Updated",
-                "Fire a test notification as if the NWS discussion was updated",
-            )
-            self._debug_menu_items["alert"] = debug_menu.Append(
-                wx.ID_ANY,
-                "Test: &Alert Notification...",
-                "Send a test alert notification (choose type and severity)",
-            )
-            self._debug_menu_items["simulate_alert"] = debug_menu.Append(
-                wx.ID_ANY,
-                "Test: &Simulate Alert Change (poll cycle)",
-                "Inject a mock alert into the next event check cycle to test the full polling path",
-            )
-            debug_menu.AppendSeparator()
-            self._debug_menu_items["diagnostics"] = debug_menu.Append(
-                wx.ID_ANY,
-                "Run Notification &Diagnostics",
-                "Run pass/fail notification system diagnostics",
-            )
-            help_menu.AppendSubMenu(debug_menu, "&Debug", "Debug and test tools")
-        help_menu.AppendSeparator()
-
-        report_issue_item = help_menu.Append(
-            wx.ID_ANY, "&Report Issue...", "Report a bug or request a feature"
-        )
-        help_menu.AppendSeparator()
-        about_item = help_menu.Append(wx.ID_ABOUT, "&About", "About AccessiWeather")
-        menu_bar.Append(help_menu, "&Help")
-
-        self.SetMenuBar(menu_bar)
-
-        # Bind menu events
-        self.Bind(wx.EVT_MENU, lambda e: self.on_settings(), settings_item)
-        self.Bind(wx.EVT_MENU, lambda e: self.app.request_exit(), exit_item)
-        self.Bind(wx.EVT_MENU, lambda e: self.on_add_location(), add_item)
-        self.Bind(wx.EVT_MENU, lambda e: self.on_edit_location(), edit_item)
-        self.Bind(wx.EVT_MENU, lambda e: self.on_remove_location(), remove_item)
-        self.Bind(wx.EVT_MENU, lambda e: self.on_refresh(), refresh_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_explain_weather(), explain_item)
-        self.Bind(wx.EVT_MENU, lambda e: self.on_view_history(), history_item)
-        self.Bind(
-            wx.EVT_MENU,
-            lambda e: self._on_precipitation_timeline(),
-            id=self._precipitation_timeline_id,
-        )
-        self.Bind(
-            wx.EVT_MENU, lambda e: self.toggle_event_center(), id=self._toggle_event_center_id
-        )
-        self.Bind(wx.EVT_MENU, lambda e: self._on_discussion(), discussion_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_aviation(), aviation_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_air_quality(), air_quality_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_uv_index(), uv_index_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_noaa_radio(), id=self._noaa_radio_id)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_weather_chat(), weather_chat_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_soundpack_manager(), soundpack_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_check_updates(), self._check_updates_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_open_user_manual(), user_manual_item)
-        if self._debug_menu_items:
-            self.Bind(
-                wx.EVT_MENU,
-                lambda e: self._on_test_discussion_notification(),
-                self._debug_menu_items["discussion"],
-            )
-            self.Bind(
-                wx.EVT_MENU,
-                lambda e: self._on_test_alert_notification(),
-                self._debug_menu_items["alert"],
-            )
-            self.Bind(
-                wx.EVT_MENU,
-                lambda e: self._on_test_notifications(),
-                self._debug_menu_items["diagnostics"],
-            )
-            self.Bind(
-                wx.EVT_MENU,
-                lambda e: self._on_debug_simulate_alert(),
-                self._debug_menu_items["simulate_alert"],
-            )
-        self.Bind(wx.EVT_MENU, lambda e: self._on_report_issue(), report_issue_item)
-        self.Bind(wx.EVT_MENU, lambda e: self._on_about(), about_item)
-        MainWindow._update_precipitation_timeline_menu_state(self)
-
-    def _on_location_changed(self, event) -> None:
-        """Handle location selection change with debounce for rapid switching."""
-        selected = self.location_dropdown.GetStringSelection()
-        if not selected:
-            return
-
-        logger.info(f"Location changed to: {selected}")
-        # Clear lifecycle labels when switching locations so stale labels never bleed across
-        self._alert_lifecycle_labels = {}
-
-        # --- All Locations special case ---
-        if selected == ALL_LOCATIONS_SENTINEL:
-            self._all_locations_active = True
-            self._update_title_for_location(ALL_LOCATIONS_SENTINEL)
-            MainWindow._safe_update_forecast_products_button_state(self)
-            MainWindow._update_precipitation_timeline_menu_state(self)
-            # Increment generation to invalidate any in-flight fetches for the previous location
-            self._fetch_generation += 1
-            # Use CallAfter so the summary renders after any already-queued wx.CallAfter
-            # callbacks (e.g. a just-completed fetch posting _on_weather_data_received) are
-            # drained first, preventing stale data from overwriting the summary.
-            wx.CallAfter(self._show_all_locations_summary)
-            self.app.run_async(self._fetch_all_locations_data())
-            return
-
-        # Switching away from All Locations view → clear the flag and stored data.
-        self._all_locations_active = False
-        self._all_locations_alerts_data = []
-        self._last_single_location_name = selected
-        self._set_forecast_sections_visible(True)
-        self._update_title_for_location(selected)
-
-        self._set_current_location(selected)
-        MainWindow._safe_update_forecast_products_button_state(self)
-
-        # Show cached data instantly if available
-        location = self.app.config_manager.get_current_location()
-        if location and hasattr(self.app, "weather_client") and self.app.weather_client:
-            cached = self.app.weather_client.get_cached_weather(location)
-            if cached and cached.has_any_data():
-                logger.info(f"Showing cached data for {selected} while refreshing")
-                self._on_weather_data_received(cached)
-
-        # Fire an immediate alert/event check for the new location (lightweight)
-        if location and hasattr(self.app, "weather_client") and self.app.weather_client:
-            self.refresh_notification_events_async()
-
-        # Debounce: cancel pending fetch, wait 500ms before fetching
-        if hasattr(self, "_location_debounce_timer") and self._location_debounce_timer.IsRunning():
-            self._location_debounce_timer.Stop()
-
-        if not hasattr(self, "_location_debounce_timer"):
-            self._location_debounce_timer = wx.Timer(self)
-            self.Bind(
-                wx.EVT_TIMER, self._on_debounced_location_fetch, self._location_debounce_timer
-            )
-
-        self._location_debounce_timer.StartOnce(500)
-
-    def _on_debounced_location_fetch(self, event) -> None:
-        """Fetch weather data after debounce period for the currently selected location."""
-        self.refresh_weather_async(force_refresh=True)
-
-    def on_add_location(self) -> None:
-        """Handle add location button click."""
-        from .dialogs import show_add_location_dialog
-
-        result = show_add_location_dialog(self, self.app)
-        if result:
-            self._populate_locations()
-            # Auto-select the newly added location so the user lands on it immediately
-            new_name = result
-            if new_name:
-                idx = self.location_dropdown.FindString(new_name)
-                if idx != wx.NOT_FOUND:
-                    self.location_dropdown.SetSelection(idx)
-                    self._set_current_location(new_name)
-            self.refresh_weather_async()
-
-    def on_edit_location(self) -> None:
-        """Handle edit location button click — currently edits marine_mode."""
-        selected = self.location_dropdown.GetStringSelection()
-        if not selected or selected == ALL_LOCATIONS_SENTINEL:
-            wx.MessageBox(
-                "Please select a specific location to edit.",
-                "No Location Selected",
-                wx.OK | wx.ICON_WARNING,
-            )
-            return
-
-        location = next(
-            (loc for loc in self.app.config_manager.get_all_locations() if loc.name == selected),
-            None,
-        )
-        if location is None:
-            return
-
-        new_marine_mode = show_edit_location_dialog(self, self.app, location)
-        if new_marine_mode is None:
-            return
-
-        self.app.config_manager.update_location_marine_mode(selected, new_marine_mode)
-        self.refresh_weather_async(force_refresh=True)
-
-    def on_remove_location(self) -> None:
-        """Handle remove location button click."""
-        selected = self.location_dropdown.GetStringSelection()
-        if not selected or selected == ALL_LOCATIONS_SENTINEL:
-            wx.MessageBox(
-                "Please select a specific location to remove.",
-                "No Location Selected",
-                wx.OK | wx.ICON_WARNING,
-            )
-            return
-
-        # Don't allow removing the last location
-        locations = self.app.config_manager.get_all_locations()
-        if len(locations) <= 1:
-            wx.MessageBox(
-                "Cannot remove the last location. Add another location first.",
-                "Cannot Remove",
-                wx.OK | wx.ICON_WARNING,
-            )
-            return
-
-        # Confirm removal
-        result = wx.MessageBox(
-            f"Are you sure you want to remove '{selected}'?",
-            "Confirm Removal",
-            wx.YES_NO | wx.ICON_QUESTION,
-        )
-        if result == wx.YES:
-            self.app.config_manager.remove_location(selected)
-            self._populate_locations()
-            self.refresh_weather_async()
-
-    def on_refresh(self) -> None:
-        """Handle refresh button click."""
-        self.refresh_weather_async(force_refresh=True)
-
-    def on_settings(self) -> None:
-        """Handle settings button click."""
-        self.open_settings()
-
-    def open_settings(self, tab: str | None = None) -> None:
-        """
-        Open the settings dialog, optionally to a specific tab.
-
-        Args:
-            tab: Optional tab name to switch to (e.g., 'Updates', 'General').
-
-        """
-        from .dialogs import show_settings_dialog
-
-        if show_settings_dialog(self, self.app, tab=tab):
-            self.app.refresh_runtime_settings()
-            # Update menu label in case update channel changed
-            self.update_check_updates_menu_label()
-            # Immediately refresh weather so source/key changes take effect
-            self.refresh_weather_async(force_refresh=True)
-
-    def on_view_history(self) -> None:
-        """View weather history comparison."""
-        from .dialogs import show_weather_history_dialog
-
-        show_weather_history_dialog(self, self.app)
-
-    def _on_precipitation_timeline(self) -> None:
-        """View Pirate Weather minutely precipitation guidance."""
-        from .dialogs import show_precipitation_timeline_dialog
-
-        show_precipitation_timeline_dialog(self, self.app)
-
-    def _on_explain_weather(self) -> None:
-        """Get AI explanation of current weather."""
-        from .dialogs import show_explanation_dialog
-
-        show_explanation_dialog(self, self.app)
-
-    def _on_discussion(self) -> None:
-        """Route to Nationwide discussion view or the per-location Forecast Products dialog."""
-        current = self.app.config_manager.get_current_location()
-        if current and current.name == "Nationwide":
-            from .dialogs.nationwide_discussion_dialog import NationwideDiscussionDialog
-
-            dlg = NationwideDiscussionDialog(parent=self, service=self._get_discussion_service())
-            dlg.ShowModal()
-            dlg.Destroy()
-        else:
-            self._on_forecast_products()
-
-    def _on_forecast_products(self) -> None:
-        """Open the Forecast Products dialog (AFD + HWO + SPS) for the active location."""
-        current = self.app.config_manager.get_current_location()
-        if current is None:
-            wx.MessageBox(
-                "Please select a location first.",
-                "No Location Selected",
-                wx.OK | wx.ICON_WARNING,
-            )
-            return
-
-        from .dialogs.forecast_products_dialog import show_forecast_products_dialog
-
-        service = self._get_forecast_product_service()
-        ai_explainer = getattr(self.app, "ai_explainer", None)
-        show_forecast_products_dialog(self, current, service, ai_explainer, app=self.app)
-
-    def _safe_update_forecast_products_button_state(self) -> None:
-        """
-        Defensive wrapper around :meth:`_update_forecast_products_button_state`.
-
-        Existing tests in ``test_all_locations_view.py`` / ``test_coverage_gaps.py``
-        drive ``MainWindow._on_location_changed`` through plain-Python stub
-        instances that deliberately omit UI-widget attributes. Wrapping the
-        toggle keeps those tests green without requiring them to add the
-        new widget to every fixture.
-        """
-        try:
-            if hasattr(self, "discussion_button") and hasattr(
-                self, "forecast_products_us_only_label"
-            ):
-                MainWindow._update_forecast_products_button_state(self)
-        except Exception:  # noqa: BLE001
-            logger.debug("Forecast products button state update skipped", exc_info=True)
-
-    def _update_forecast_products_button_state(self) -> None:
-        """
-        Enable/disable the Forecast Products button based on country.
-
-        Non-US locations disable the button and reveal the adjacent
-        "NWS products are US-only" StaticText so screen readers announce the
-        reason. US locations (and the Nationwide entry) re-enable the button
-        and hide the label.
-        """
-        try:
-            current = self.app.config_manager.get_current_location()
-        except Exception:  # noqa: BLE001
-            current = None
-
-        is_us = True  # default: don't needlessly disable
-        if current is not None and current.name != "Nationwide":
-            country = getattr(current, "country_code", None)
-            if country:
-                is_us = country.upper() == "US"
-
-        if is_us:
-            self.discussion_button.Enable()
-            self.forecast_products_us_only_label.Hide()
-        else:
-            self.discussion_button.Disable()
-            self.forecast_products_us_only_label.Show()
-
-    def _get_forecast_product_service(self):
-        """Get or lazily build the shared ForecastProductService instance."""
-        existing = getattr(self, "_forecast_product_service", None)
-        if existing is not None:
-            return existing
-
-        from ..cache import Cache
-        from ..services.forecast_product_service import ForecastProductService
-
-        # Prefer a cache shared with the rest of the app when one exists;
-        # fall back to an owned instance. Keeps tests from having to wire
-        # the full app graph just to construct the dialog.
-        cache = getattr(self.app, "cache", None) or Cache()
-        self._forecast_product_service = ForecastProductService(cache)
-        return self._forecast_product_service
-
-    def _on_aviation(self) -> None:
-        """View aviation weather."""
-        from .dialogs import show_aviation_dialog
-
-        show_aviation_dialog(self, self.app)
-
-    def _on_air_quality(self) -> None:
-        """View air quality information."""
-        from .dialogs import show_air_quality_dialog
-
-        show_air_quality_dialog(self, self.app)
-
-    def _on_uv_index(self) -> None:
-        """View UV index information."""
-        from .dialogs import show_uv_index_dialog
-
-        show_uv_index_dialog(self, self.app)
-
-    def _on_noaa_radio(self) -> None:
-        """Open NOAA Weather Radio dialog."""
-        location = self.app.config_manager.get_current_location()
-        if not location:
-            wx.MessageBox(
-                "Please select a location first.",
-                "No Location Selected",
-                wx.OK | wx.ICON_WARNING,
-            )
-            return
-
-        from .dialogs import show_noaa_radio_dialog
-
-        show_noaa_radio_dialog(self, location.latitude, location.longitude)
-
-    def _on_weather_chat(self) -> None:
-        """Open Weather Assistant dialog."""
-        from .dialogs import show_weather_assistant_dialog
-
-        show_weather_assistant_dialog(self, self.app)
-
-    def _on_soundpack_manager(self) -> None:
-        """Open the soundpack manager dialog."""
-        from .dialogs import show_soundpack_manager_dialog
-
-        show_soundpack_manager_dialog(self, self.app)
-
-    def _get_update_channel(self) -> str:
-        """Get the configured update channel from settings."""
-        try:
-            settings = self.app.config_manager.get_settings()
-            return getattr(settings, "update_channel", "stable")
-        except Exception:
-            return "stable"
-
-    def _on_check_updates(self) -> None:
-        """Check for updates from the Help menu."""
-        import asyncio
-
-        from ..services.simple_update import UpdateService, parse_nightly_date
-
-        # Skip update checks when running from source
-        if not is_compiled_runtime():
-            wx.MessageBox(
-                "Update checking is only available in installed builds.\n"
-                "You're running from source — use git pull to update.",
-                "Running from Source",
-                wx.OK | wx.ICON_INFORMATION,
-            )
-            return
-
-        channel = self._get_update_channel()
-        current_version = getattr(self.app, "version", "0.0.0")
-        build_tag = getattr(self.app, "build_tag", None)
-        current_nightly_date = parse_nightly_date(build_tag) if build_tag else None
-        # Show nightly date as the display version when running a nightly build
-        display_version = current_nightly_date if current_nightly_date else current_version
-
-        # Show checking status
-        wx.BeginBusyCursor()
-
-        def do_check():
-            try:
-
-                async def check():
-                    service = UpdateService("AccessiWeather")
-                    try:
-                        return await service.check_for_updates(
-                            current_version=current_version,
-                            current_nightly_date=current_nightly_date,
-                            channel=channel,
-                        )
-                    finally:
-                        await service.close()
-
-                update_info = asyncio.run(check())
-                wx.CallAfter(wx.EndBusyCursor)
-
-                if update_info is None:
-                    # No update available - show appropriate message
-                    if current_nightly_date and channel == "stable":
-                        msg = (
-                            f"You're on nightly ({current_nightly_date}).\n"
-                            "No newer stable release available."
-                        )
-                    elif current_nightly_date:
-                        msg = f"You're on the latest nightly ({current_nightly_date})."
-                    else:
-                        msg = f"You're up to date ({display_version})."
-
-                    wx.CallAfter(
-                        wx.MessageBox,
-                        msg,
-                        "No Updates Available",
-                        wx.OK | wx.ICON_INFORMATION,
-                    )
-                else:
-                    # Update available — show changelog dialog
-                    channel_label = "Nightly" if update_info.is_nightly else "Stable"
-
-                    def prompt():
-                        from .dialogs.update_dialog import UpdateAvailableDialog
-
-                        dlg = UpdateAvailableDialog(
-                            parent=self,
-                            current_version=display_version,
-                            new_version=update_info.version,
-                            channel_label=channel_label,
-                            release_notes=update_info.release_notes,
-                        )
-                        result = dlg.ShowModal()
-                        dlg.Destroy()
-                        if result == wx.ID_OK:
-                            self.app._download_and_apply_update(update_info)
-
-                    wx.CallAfter(prompt)
-
-            except Exception as e:
-                wx.CallAfter(wx.EndBusyCursor)
-                wx.CallAfter(
-                    wx.MessageBox,
-                    f"Failed to check for updates:\n{e}",
-                    "Update Check Failed",
-                    wx.OK | wx.ICON_ERROR,
-                )
-
-        import threading
-
-        thread = threading.Thread(target=do_check, daemon=True)
-        thread.start()
-
-    def _on_test_discussion_notification(self) -> None:
-        """Fire a real test notification simulating an NWS discussion update."""
-        from ..notifications.toast_notifier import SafeDesktopNotifier
-
-        try:
-            settings = self.app.config_manager.get_settings()
-        except Exception:
-            from ..models import AppSettings
-
-            settings = AppSettings()
-
-        notifier = getattr(self.app, "notifier", None) or SafeDesktopNotifier(
-            app_name="AccessiWeather",
-            sound_enabled=bool(getattr(settings, "sound_enabled", True)),
-            soundpack=getattr(settings, "sound_pack", "default"),
-            muted_sound_events=getattr(settings, "muted_sound_events", []),
-        )
-        from ..notification_activation import (
-            NotificationActivationRequest,
-            serialize_activation_request,
-        )
-
-        sent = notifier.send_notification(
-            title="NWS Discussion Updated",
-            message="The Area Forecast Discussion for your location has been updated. "
-            "This is a debug test notification.",
-            timeout=10,
-            sound_candidates=["discussion_update", "notify"],
-            play_sound=True,
-            activation_arguments=serialize_activation_request(
-                NotificationActivationRequest(kind="discussion")
-            ),
-        )
-        if not sent:
-            wx.MessageBox(
-                "Discussion notification could not be sent.\n"
-                "Check that desktop notifications are enabled on your system.",
-                "Debug: Discussion Notification",
-                wx.OK | wx.ICON_WARNING,
-            )
-
-    def _on_test_alert_notification(self) -> None:
-        """Open the alert notification test dialog."""
-        from .dialogs.debug_alert_dialog import DebugAlertDialog
-
-        dlg = DebugAlertDialog(self, self.app)
-        dlg.ShowModal()
-        dlg.Destroy()
-
-    def _on_debug_simulate_alert(self) -> None:
-        """Inject a mock alert into the event check cycle to test the full polling path."""
-        from ..alert_lifecycle import AlertChange, AlertChangeKind, AlertLifecycleDiff
-        from ..models.alerts import WeatherAlert
-        from ..models.weather import WeatherAlerts, WeatherData
-
-        app = self.app
-        if not hasattr(app, "weather_client") or not app.weather_client:
-            wx.MessageBox("Weather client not ready.", "Debug", wx.OK | wx.ICON_WARNING)
-            return
-
-        location = app.config_manager.get_current_location()
-        if not location:
-            wx.MessageBox("No current location.", "Debug", wx.OK | wx.ICON_WARNING)
-            return
-
-        fake_alert = WeatherAlert(
-            title="Tornado Warning",
-            description=(
-                "This is a simulated alert injected via the debug menu "
-                "to test the event polling notification path."
-            ),
-            severity="Extreme",
-            urgency="Immediate",
-            certainty="Observed",
-            event="Tornado Warning",
-            headline="DEBUG: Simulated Tornado Warning for testing",
-            areas=["Test County"],
-            id="debug-simulate-001",
-            message_type="Alert",
-        )
-        fake_alerts = WeatherAlerts(alerts=[fake_alert])
-        fake_change = AlertChange(
-            kind=AlertChangeKind.NEW,
-            alert=fake_alert,
-            alert_id=fake_alert.get_unique_id(),
-            title=fake_alert.title,
-        )
-        fake_diff = AlertLifecycleDiff(new_alerts=[fake_change])
-
-        mock_data = WeatherData(location=location)
-        mock_data.alerts = fake_alerts
-        mock_data.alert_lifecycle_diff = fake_diff
-
-        self._on_notification_event_data_received(mock_data)
-
-    def _on_test_notifications(self) -> None:
-        """Run notification tests and show pass/fail results."""
-        from ..notifications.notification_test import run_notification_test
-
-        results = run_notification_test(self.app)
-        result_keys = [
-            "safe_desktop_notifier",
-            "alert_notification_system",
-            "discussion_update_path",
-        ]
-        lines = ["Notification test results:"]
-        for key in result_keys:
-            result = results.get(key, {})
-            label = key.replace("_", " ").title()
-            status = "PASS" if result.get("passed") else "FAIL"
-            detail = result.get("message", "")
-            lines.append(f"- {label}: {status}")
-            if detail:
-                lines.append(f"  {detail}")
-        summary = f"Summary: {results.get('passed_count', 0)}/{results.get('total_count', len(result_keys))} passed"
-        lines.append("")
-        lines.append(summary)
-        wx.MessageBox(
-            "\n".join(lines),
-            "Notification Test Results",
-            wx.OK | (wx.ICON_INFORMATION if results.get("all_passed") else wx.ICON_WARNING),
-        )
-
-    def update_check_updates_menu_label(self) -> None:
-        """Update the Check for Updates menu item label with current channel."""
-        channel = self._get_update_channel()
-        self._check_updates_item.SetItemLabel(f"Check for &Updates ({channel.title()})...")
-
-    def _on_report_issue(self) -> None:
-        """Open the report issue dialog."""
-        from .dialogs.report_issue_dialog import ReportIssueDialog
-
-        dialog = ReportIssueDialog(self)
-        dialog.ShowModal()
-        dialog.Destroy()
-
-    def _on_open_user_manual(self) -> None:
-        """Open the bundled user manual or fall back to the online manual."""
-        if open_user_manual():
-            return
-
-        wx.MessageBox(
-            "AccessiWeather could not open the user manual.",
-            "User Manual Unavailable",
-            wx.OK | wx.ICON_ERROR,
-        )
-
-    def _on_about(self) -> None:
-        """Show about dialog."""
-        from accessiweather import __version__
-
-        portable = bool(getattr(self.app, "_portable_mode", False))
-        mode_label = "Portable" if portable else "Installed"
-        config_path = (
-            str(self.app.config_manager.config_dir) if self.app.config_manager else "unknown"
-        )
-
-        wx.MessageBox(
-            f"AccessiWeather v{__version__}\n\n"
-            "An accessible weather application with NOAA and Open-Meteo support.\n\n"
-            "Built with wxPython for screen reader compatibility.\n\n"
-            f"Mode: {mode_label}\n"
-            f"Config path: {config_path}\n\n"
-            "https://github.com/Orinks/AccessiWeather",
-            "About AccessiWeather",
-            wx.OK | wx.ICON_INFORMATION,
-        )
-
-    def _on_view_alert(self, event) -> None:
-        """Handle view alert button click."""
-        selected = self.alerts_list.GetSelection()
-        if selected != wx.NOT_FOUND:
-            self._show_alert_details(selected)
-
-    def _on_close(self, event) -> None:
-        """Handle window close event."""
-        # Check if minimize to tray is enabled
-        if self._should_minimize_to_tray():
-            self._minimize_to_tray()
-            event.Veto()  # Prevent the window from being destroyed
-            return
-
-        # Otherwise, exit the application
-        self._announcer.shutdown()
-        self.app.request_exit()
-
-    def _on_iconize(self, event) -> None:
-        """Handle window iconize (minimize) event."""
-        # Check if minimize to tray is enabled and window is being minimized
-        if event.IsIconized() and self._should_minimize_to_tray():
-            # Use CallAfter to let the iconize event complete before hiding
-            wx.CallAfter(self._minimize_to_tray)
-            return
-        event.Skip()  # Allow normal minimize behavior
-
-    def _setup_escape_accelerator(self) -> None:
-        """
-        Set up accelerator table for Escape key and menu shortcuts.
-
-        wxPython's SetAcceleratorTable replaces the implicit menu accelerators,
-        so we must include all Ctrl+ shortcuts here alongside Escape.
-        """
-        self._escape_id = wx.NewIdRef()
-        entries = [
-            (wx.ACCEL_NORMAL, wx.WXK_ESCAPE, self._escape_id),
-            (wx.ACCEL_NORMAL, wx.WXK_F5, wx.ID_REFRESH),
-        ]
-        # Re-register all Ctrl+ menu accelerators (SetAcceleratorTable replaces them)
-        ctrl_shortcuts = [
-            (wx.ACCEL_CTRL, "S", "_settings_id"),
-            (wx.ACCEL_CTRL, "Q", "_exit_id"),
-            (wx.ACCEL_CTRL, "L", "_add_location_id"),
-            (wx.ACCEL_CTRL, "D", "_remove_location_id"),
-            (wx.ACCEL_CTRL, "E", "_explain_id"),
-            (wx.ACCEL_CTRL, "H", "_history_id"),
-            (wx.ACCEL_CTRL | wx.ACCEL_SHIFT, "R", "_noaa_radio_id"),
-            (wx.ACCEL_CTRL, "T", "_weather_chat_id"),
-        ]
-        for flags, key, attr in ctrl_shortcuts:
-            if hasattr(self, attr):
-                entries.append((flags, ord(key), getattr(self, attr)))
-        accel_tbl = wx.AcceleratorTable(entries)
-        self.SetAcceleratorTable(accel_tbl)
-        # Bind the escape action
-        self.Bind(wx.EVT_MENU, self._on_escape_pressed, id=self._escape_id)
-
-    def _on_escape_pressed(self, event) -> None:
-        """Handle Escape key press via accelerator."""
-        if self._should_minimize_to_tray():
-            self._minimize_to_tray()
-        # If minimize to tray is disabled, do nothing (Escape has no effect)
-
-    def _should_minimize_to_tray(self) -> bool:
-        """
-        Check if minimize to tray is enabled in settings.
-
-        Returns:
-            True if minimize to tray is enabled, False otherwise
-
-        """
-        try:
-            settings = self.app.config_manager.get_settings()
-            return bool(getattr(settings, "minimize_to_tray", False))
-        except Exception:
-            return False
-
-    def _minimize_to_tray(self) -> None:
-        """Minimize the window to the system tray."""
-        try:
-            self.Iconize(False)  # Restore from iconized state first
-            self.Hide()
-            logger.debug("Window minimized to system tray")
-        except Exception as e:
-            logger.error(f"Failed to minimize to tray: {e}")
-
-    def _set_current_location(self, location_name: str) -> None:
-        """Set the current location and persist to config."""
-        try:
-            # set_current_location expects a string (location name), not a Location object
-            result = self.app.config_manager.set_current_location(location_name)
-            if not result:
-                logger.error(
-                    f"Failed to set current location '{location_name}': "
-                    "location not found or save failed"
-                )
-            else:
-                logger.info(f"Current location set and saved: {location_name}")
-        except Exception as e:
-            logger.error(f"Failed to set current location: {e}")
-
-    def refresh_weather_async(self, force_refresh: bool = False) -> None:
-        """
-        Refresh weather data asynchronously.
-
-        When the All Locations view is active the summary is rebuilt from
-        whatever is currently in the cache — no new network requests are made.
-        """
-        # All Locations view: fetch fresh data for all locations sequentially then re-render.
-        if getattr(self, "_all_locations_active", False):
-            self.refresh_button.Disable()
-            self.app.run_async(self._fetch_all_locations_data())
-            return
-
-        if self.app.is_updating and not force_refresh:
-            logger.debug("Already updating, skipping refresh")
-            return
-
-        # Increment generation to invalidate any in-flight fetches
-        self._fetch_generation += 1
-        self.app.is_updating = True
-        self.refresh_button.Disable()
-
-        # Run async weather fetch with current generation
-        generation = self._fetch_generation
-        self.app.run_async(
-            self._fetch_weather_data(force_refresh=force_refresh, generation=generation)
-        )
-
-    def refresh_notification_events_async(self) -> None:
-        """Run a lightweight event check without refreshing the full weather UI."""
-        main_window_notification_events.refresh_notification_events_async(self)
-
-    async def _fetch_notification_event_data(self) -> None:
-        """Fetch only the lightweight data needed for notifications."""
-        await main_window_notification_events.fetch_notification_event_data(self)
-
-    async def _fetch_weather_data(self, force_refresh: bool = False, generation: int = 0) -> None:
-        """Fetch weather data in background."""
-        try:
-            location = self.app.config_manager.get_current_location()
-            if not location:
-                wx.CallAfter(self._on_weather_error, "No location selected")
-                return
-
-            # For Nationwide location, fetch discussion summaries instead of weather
-            if location.name == "Nationwide":
-                wx.CallAfter(
-                    self.current_conditions.SetValue,
-                    "Fetching nationwide weather discussions from NWS, SPC, NHC, and CPC...\n"
-                    "This may take a moment.",
-                )
-                wx.CallAfter(self._set_forecast_sections, "", "")
-                await self._fetch_nationwide_discussions(generation)
-                return
-
-            # Fetch weather data - pass the Location object directly
-            # force_refresh=True bypasses cache (used when switching locations)
-            weather_data = await self.app.weather_client.get_weather_data(
-                location, force_refresh=force_refresh
-            )
-
-            # Only update UI if this fetch is still current (not superseded by a newer one)
-            if generation != self._fetch_generation:
-                logger.debug(
-                    f"Discarding stale fetch for {location.name} "
-                    f"(gen {generation} < {self._fetch_generation})"
-                )
-                return
-
-            # Update UI on main thread
-            wx.CallAfter(self._on_weather_data_received, weather_data)
-
-            # Pre-warm NWS text products (AFD/HWO/SPS) for the active location
-            # so the Forecast Products dialog and Unit 10/11 notification
-            # checks see fresh data without issuing an on-demand fetch.
-            await self._pre_warm_products_for_location(location)
-
-            # Pre-warm cache for other locations in background (non-blocking)
-            if not force_refresh:
-                await self._pre_warm_other_locations(location)
-
-        except Exception as e:
-            logger.error(f"Failed to fetch weather data: {e}")
-            wx.CallAfter(self._on_weather_error, str(e))
-
-    def _get_discussion_service(self):
-        """Get or create the shared NationalDiscussionService instance."""
-        if not hasattr(self, "_discussion_service") or self._discussion_service is None:
-            from ..services.national_discussion_service import NationalDiscussionService
-
-            self._discussion_service = NationalDiscussionService()
-        return self._discussion_service
-
-    async def _fetch_nationwide_discussions(self, generation: int) -> None:
-        """Fetch nationwide discussion summaries and display in weather fields."""
-        import asyncio
-
-        try:
-            service = self._get_discussion_service()
-            # Run synchronous fetch in thread to avoid blocking
-            discussions = await asyncio.to_thread(service.fetch_all_discussions)
-
-            if generation != self._fetch_generation:
-                return
-
-            # Build current conditions text (short range + SPC day 1)
-            current_parts = ["=== Nationwide Weather Summary ===\n"]
-            wpc = discussions.get("wpc", {})
-            if wpc.get("short_range", {}).get("text"):
-                current_parts.append(
-                    f"--- {wpc['short_range']['title']} ---\n{wpc['short_range']['text']}\n"
-                )
-            spc = discussions.get("spc", {})
-            if spc.get("day1", {}).get("text"):
-                current_parts.append(f"--- {spc['day1']['title']} ---\n{spc['day1']['text']}\n")
-            current_text = "\n".join(current_parts)
-
-            # Build forecast text (extended + CPC outlooks)
-            forecast_parts = ["=== Extended Outlook ===\n"]
-            if wpc.get("extended", {}).get("text"):
-                forecast_parts.append(
-                    f"--- {wpc['extended']['title']} ---\n{wpc['extended']['text']}\n"
-                )
-            cpc = discussions.get("cpc", {})
-            if cpc.get("outlook", {}).get("text"):
-                forecast_parts.append(
-                    f"--- {cpc['outlook']['title']} ---\n{cpc['outlook']['text']}\n"
-                )
-            forecast_text = "\n".join(forecast_parts)
-
-            wx.CallAfter(self._on_nationwide_data_received, current_text, forecast_text)
-
-        except Exception as e:
-            logger.error(f"Failed to fetch nationwide discussions: {e}")
-            wx.CallAfter(self._on_weather_error, str(e))
-
-    def _on_nationwide_data_received(self, current_text: str, forecast_text: str) -> None:
-        """Handle received nationwide discussion data (called on main thread)."""
-        try:
-            self.current_conditions.SetValue(current_text)
-            self._set_forecast_sections(forecast_text, "")
-            self.stale_warning_label.SetLabel("")
-        except Exception as e:
-            logger.error(f"Error updating nationwide display: {e}")
-        finally:
-            self.app.is_updating = False
-            self.refresh_button.Enable()
-
-    async def _fetch_all_locations_data(self) -> None:
-        """Fetch fresh weather data for all saved locations sequentially, then re-render summary."""
-        try:
-            all_locations = [
-                loc
-                for loc in self.app.config_manager.get_all_locations()
-                if loc.name != "Nationwide"
-            ]
-            if all_locations and self.app.weather_client:
-                await self.app.weather_client.pre_warm_batch(all_locations)
-        except Exception as e:
-            logger.error(f"Failed to refresh all locations: {e}")
-        finally:
-            wx.CallAfter(self._on_all_locations_refresh_complete)
-
-    def _on_all_locations_refresh_complete(self) -> None:
-        """Handle completion of all-locations background refresh on the main thread."""
-        if getattr(self, "_all_locations_active", False):
-            self._show_all_locations_summary()
-        self.refresh_button.Enable()
-        self.app.is_updating = False
-
-    async def _pre_warm_other_locations(self, current_location: Location) -> None:
-        """Pre-warm cache for non-current locations so switching is instant."""
-        try:
-            all_locations = self.app.config_manager.get_all_locations()
-            uncached = [
-                loc
-                for loc in all_locations
-                if loc.name != current_location.name
-                and not (
-                    self.app.weather_client.get_cached_weather(loc)
-                    and self.app.weather_client.get_cached_weather(loc).has_any_data()
-                )
-            ]
-            if uncached:
-                logger.debug(f"Pre-warming cache for {len(uncached)} locations")
-                await self.app.weather_client.pre_warm_batch(uncached)
-
-            # Pre-warm NWS text products (AFD/HWO/SPS) for every non-active
-            # saved US location. Failure isolation is per-(product, location):
-            # one failure never cascades to other products or other locations.
-            for loc in all_locations:
-                if loc.name == current_location.name:
-                    continue
-                await self._pre_warm_products_for_location(loc)
-        except Exception as e:
-            logger.debug(f"Cache pre-warm failed (non-critical): {e}")
-
-    async def _pre_warm_products_for_location(self, location: Location) -> None:
-        """
-        Pre-warm AFD/HWO/SPS caches for a single location.
-
-        Non-US locations and US locations without a populated ``cwa_office``
-        are skipped. Each product fetch is wrapped in its own try/except so
-        one failure (e.g. an NWS 404 for HWO) never prevents the next product
-        type or subsequent locations from being pre-warmed.
-        """
-        # Nationwide is a synthetic entry with no real CWA — skip it cheaply.
-        if location.name == "Nationwide":
-            return
-
-        # Local import avoids a hard dependency on services at module-import
-        # time for the stripped-down test fixtures that stub MainWindow.
-        from ..services.zone_enrichment_service import _is_us_location
-        from ..weather_client_nws import TextProductFetchError
-
-        if not _is_us_location(location):
-            return
-        if not getattr(location, "cwa_office", None):
-            return
-
-        try:
-            service = self._get_forecast_product_service()
-        except Exception:  # noqa: BLE001
-            logger.debug("ForecastProductService unavailable for pre-warm", exc_info=True)
-            return
-
-        for product_type in ("AFD", "HWO", "SPS"):
-            try:
-                await service.get(product_type, location.cwa_office)
-            except TextProductFetchError:
-                logger.debug(
-                    "Pre-warm %s for %s (%s) failed",
-                    product_type,
-                    location.name,
-                    location.cwa_office,
-                )
-            except Exception:  # noqa: BLE001
-                logger.debug(
-                    "Unexpected pre-warm failure for %s at %s",
-                    product_type,
-                    location.name,
-                    exc_info=True,
-                )
-
-    def _on_weather_data_received(self, weather_data) -> None:
-        """Handle received weather data (called on main thread)."""
-        # Guard: if we switched to All Locations view, ignore stale single-location data.
-        if getattr(self, "_all_locations_active", False):
-            logger.debug("Ignoring stale weather data received while All Locations view is active")
-            return
-        try:
-            self.app.current_weather_data = weather_data
-            MainWindow._update_precipitation_timeline_menu_state(self, weather_data)
-
-            # Use presenter to create formatted presentation
-            presentation = self.app.presenter.present(weather_data)
-
-            # Update current conditions (with data source attribution appended)
-            current_text = ""
-            if presentation.current_conditions:
-                current_text = presentation.current_conditions.fallback_text
-            else:
-                current_text = "No current conditions available."
-
-            # Append data source attribution to current conditions for screen reader accessibility
-            if presentation.source_attribution and presentation.source_attribution.summary_text:
-                current_text += f"\n\n{presentation.source_attribution.summary_text}"
-
-            self.current_conditions.SetValue(current_text)
-
-            # Update stale/cached data warning
-            if presentation.status_messages:
-                warning_text = " ".join(presentation.status_messages)
-                self.stale_warning_label.SetLabel(warning_text)
-            else:
-                self.stale_warning_label.SetLabel("")
-
-            # Update forecast
-            if presentation.forecast:
-                daily_sections = [presentation.forecast.daily_section_text]
-                if presentation.forecast.marine_section_text:
-                    daily_sections.append(presentation.forecast.marine_section_text)
-                daily_text = "\n\n".join(section for section in daily_sections if section).rstrip()
-                daily_text = daily_text or "No daily forecast available."
-                hourly_text = (
-                    presentation.forecast.hourly_section_text or "No hourly forecast available."
-                )
-                self._set_forecast_sections(daily_text, hourly_text)
-                if presentation.forecast.mobility_briefing:
-                    self.append_event_center_entry(
-                        presentation.forecast.mobility_briefing,
-                        category="Briefing",
-                    )
-            else:
-                self._set_forecast_sections(
-                    "No daily forecast available.", "No hourly forecast available."
-                )
-
-            # Update lifecycle label map from the current active alerts, then refresh the alerts list.
-            if weather_data.alerts is not None:
-                from accessiweather.alert_lifecycle import compute_lifecycle_labels
-
-                active_alerts = weather_data.alerts.get_active_alerts()
-                self._alert_lifecycle_labels = compute_lifecycle_labels(active_alerts)
-
-            # Update alerts
-            self._update_alerts(weather_data.alerts, self._alert_lifecycle_labels)
-
-            # Process alert notifications on full refresh too (AlertManager deduplicates
-            # so the lightweight event poll won't re-notify for the same alerts).
-            if (
-                weather_data.alerts
-                and weather_data.alerts.has_alerts()
-                and self.app.alert_notification_system
-            ):
-                active_alerts = weather_data.alerts.get_active_alerts()
-                logger.info(
-                    "[notify-ui] full refresh scheduling alert processing for %d active alert(s): %s",
-                    len(active_alerts),
-                    [
-                        {
-                            "id": alert.get_unique_id(),
-                            "event": alert.event,
-                            "severity": alert.severity,
-                        }
-                        for alert in active_alerts
-                    ],
-                )
-                self.app.run_async(
-                    self.app.alert_notification_system.process_and_notify(weather_data.alerts)
-                )
-
-            location = self.app.config_manager.get_current_location()
-            location_name = location.name if location else "Unknown"
-
-            # Update system tray tooltip with current weather
-            self.app.update_tray_tooltip(weather_data, location_name)
-
-            # Process notification events (AFD updates, severe risk changes)
-            self._process_notification_events(weather_data)
-
-            # Surface the refresh time in the status bar so users have a
-            # passive indicator of data freshness without needing to re-read
-            # any panels.  Silent (no screen-reader announcement).
-            self._set_last_updated_status()
-
-            # Play the weather-updated sound on successful refresh.
-            try:
-                settings = self.app.config_manager.get_settings()
-                if getattr(settings, "sound_enabled", True):
-                    from accessiweather.notifications.sound_player import play_data_updated_sound
-
-                    sound_pack = getattr(settings, "sound_pack", "default")
-                    muted_events = getattr(settings, "muted_sound_events", [])
-                    play_data_updated_sound(sound_pack, muted_events=muted_events)
-            except Exception as sound_exc:
-                logger.debug(f"Failed to play weather-updated sound: {sound_exc}")
-
-        except Exception as e:
-            logger.error(f"Failed to update weather display: {e}")
-            self.set_status(f"Error updating display: {e}")
-
-        finally:
-            self.app.is_updating = False
-            self.refresh_button.Enable()
-
-    def _set_forecast_sections(self, daily_text: str, hourly_text: str) -> None:
-        """Update the daily and hourly forecast controls together."""
-        self.daily_forecast_display.SetValue(daily_text)
-        self.hourly_forecast_display.SetValue(hourly_text)
-
-    def append_event_center_entry(self, text: str, *, category: str | None = None) -> None:
-        """Append a timestamped reviewable line to the Event Center."""
-        if not text:
-            return
-        timestamp = datetime.now().strftime("%I:%M %p").lstrip("0")
-        prefix = f"{category}: " if category else ""
-        entry = f"[{timestamp}] {prefix}{text}\n"
-        self.event_center_display.AppendText(entry)
-
-    def toggle_event_center(self) -> None:
-        """Show or hide the Event Center section."""
-        visible = not getattr(self, "_event_center_visible", True)
-        self._event_center_visible = visible
-        self._event_center_label.Show(visible)
-        self.event_center_display.Show(visible)
-        self.Layout()
-
-    def focus_event_center(self) -> None:
-        """Reveal and focus the Event Center section."""
-        if not getattr(self, "_event_center_visible", True):
-            self._event_center_visible = True
-            self._event_center_label.Show(True)
-            self.event_center_display.Show(True)
-            self.Layout()
-        self.event_center_display.SetFocus()
-
-    def get_visible_top_level_sections(self) -> list[tuple[str, wx.Window]]:
-        """Return the canonical visible top-level weather sections in focus order."""
-        sections: list[tuple[str, wx.Window]] = [
-            ("Location", self.location_dropdown),
-            ("Current conditions", self.current_conditions),
-            ("Hourly / near-term", self.hourly_forecast_display),
-            ("Daily forecast", self.daily_forecast_display),
-            ("Alerts", self.alerts_list),
-        ]
-        if getattr(self, "_event_center_visible", True):
-            sections.append(("Event Center", self.event_center_display))
-        return sections
-
-    def focus_section_by_number(self, number: int) -> None:
-        """Focus a canonical top-level section by its 1-based shortcut number."""
-        if number == 5 and not getattr(self, "_event_center_visible", True):
-            return
-
-        sections = self.get_visible_top_level_sections()
-        index = len(sections) - 1 if number == 5 else number
-        if 0 <= index < len(sections):
-            _label, widget = sections[index]
-            widget.SetFocus()
-
-    def cycle_section_focus(self) -> None:
-        """Move focus to the next visible top-level section, wrapping at the end."""
-        sections = self.get_visible_top_level_sections()
-        if not sections:
-            return
-
-        next_index = getattr(self, "_section_focus_index", -1) + 1
-        if next_index >= len(sections):
-            next_index = 0
-
-        self._section_focus_index = next_index
-        _label, widget = sections[next_index]
-        widget.SetFocus()
-
-    def _set_forecast_sections_visible(self, visible: bool) -> None:
-        """Show or hide the daily/hourly forecast labels and controls."""
-        for attr in (
-            "_daily_forecast_label",
-            "daily_forecast_display",
-            "_hourly_forecast_label",
-            "hourly_forecast_display",
-        ):
-            widget = getattr(self, attr, None)
-            if widget is not None:
-                widget.Show(visible)
-        sizer = self.GetSizer() if callable(getattr(self, "GetSizer", None)) else None
-        if sizer:
-            sizer.Layout()
-
-    def _on_weather_error(self, error_message: str) -> None:
-        """Handle weather fetch error (called on main thread)."""
-        self.set_status(f"Error: {error_message}")
-        self.app.is_updating = False
-        self.refresh_button.Enable()
-
-        # Play fetch_error sound on weather data fetch failure
-        try:
-            settings = self.app.config_manager.get_settings()
-            if getattr(settings, "sound_enabled", True):
-                from accessiweather.notifications.sound_player import play_fetch_error_sound
-
-                sound_pack = getattr(settings, "sound_pack", "default")
-                muted_events = getattr(settings, "muted_sound_events", [])
-                play_fetch_error_sound(sound_pack, muted_events=muted_events)
-        except Exception as sound_exc:
-            logger.debug(f"Failed to play fetch_error sound: {sound_exc}")
-
-    def _update_alerts(self, alerts, lifecycle_labels: dict[str, str] | None = None) -> None:
-        """
-        Update the alerts list.
-
-        Args:
-            alerts: WeatherAlerts, a list of WeatherAlert, or None.
-            lifecycle_labels: Optional mapping of alert_id -> label (e.g.
-                "New", "Updated", "Escalated", "Extended").  When provided,
-                the label is appended to each matching list item:
-                "Dense Fog Advisory (Moderate) (Extended)".
-
-        """
-        alert_items = []
-        alert_list = []
-
-        # Handle WeatherAlerts object or list.
-        # Prefer get_active_alerts() so expired alerts are never shown in the
-        # listbox (cached data may contain alerts that expired while cached).
-        if alerts:
-            if hasattr(alerts, "get_active_alerts"):
-                alert_list = alerts.get_active_alerts() or []
-            elif hasattr(alerts, "alerts"):
-                alert_list = alerts.alerts or []
-            elif isinstance(alerts, list):
-                alert_list = alerts
-
-        for alert in alert_list:
-            event = getattr(alert, "event", "Unknown")
-            severity = getattr(alert, "severity", "Unknown")
-            item = f"{event} ({severity})"
-            if lifecycle_labels:
-                get_uid = getattr(alert, "get_unique_id", None)
-                if callable(get_uid):
-                    label = lifecycle_labels.get(get_uid())
-                    if label:
-                        item = f"{item} ({label})"
-            alert_items.append(item)
-
-        self.alerts_list.Clear()
-        if alert_items:
-            self.alerts_list.Append(alert_items)
-
-        # Enable/disable view button based on alerts
-        if alert_items:
-            self.view_alert_button.Enable()
-        else:
-            self.view_alert_button.Disable()
-
-    def _show_alert_details(self, alert_index: int) -> None:
-        """Show details for the selected alert."""
-        # In All Locations mode alerts come from the aggregated list, not current_weather_data.
-        if getattr(self, "_all_locations_active", False):
-            data = getattr(self, "_all_locations_alerts_data", [])
-            if 0 <= alert_index < len(data):
-                _loc_name, alert = data[alert_index]
-                from .dialogs import show_alert_dialog
-
-                show_alert_dialog(self, alert, self.app.config_manager.get_settings())
-            return
-
-        if not self.app.current_weather_data or not self.app.current_weather_data.alerts:
-            return
-
-        alerts = self.app.current_weather_data.alerts
-        active = alerts.get_active_alerts()
-        if 0 <= alert_index < len(active):
-            alert = active[alert_index]
-            from .dialogs import show_alert_dialog
-
-            show_alert_dialog(self, alert, self.app.config_manager.get_settings())
-
-    def _show_all_locations_summary(self) -> None:
-        """
-        Build and display a cached-data summary for every saved location.
-
-        No network requests are made.  The daily and hourly forecast sections
-        are replaced with a short explanatory message because per-location
-        forecasts are not meaningful in this aggregate view.  Alerts from all
-        locations are collected and shown in the alerts list with the location
-        name as a prefix so screen-reader users can identify the source.
-
-        Nationwide is excluded from the summary because it is a special
-        aggregate view of its own, not an individual saved location.
-        """
-        try:
-            all_locs = [
-                loc
-                for loc in self.app.config_manager.get_all_locations()
-                if loc.name != "Nationwide"
-            ]
-        except Exception as e:
-            logger.error(f"Failed to get locations for All Locations summary: {e}")
-            all_locs = []
-
-        if not all_locs:
-            self.current_conditions.SetValue(
-                "No locations configured.\nUse the Add button or Ctrl+L to add a location."
-            )
-            self._set_forecast_sections("", "")
-            self.alerts_list.Clear()
-            self.view_alert_button.Disable()
-            MainWindow._update_precipitation_timeline_menu_state(self)
-            return
-
-        lines: list[str] = ["All Locations Summary", ""]
-        location_alerts: list[tuple[str, object]] = []
-
-        weather_client = (
-            self.app.weather_client
-            if hasattr(self.app, "weather_client") and self.app.weather_client
-            else None
-        )
-
-        for loc in all_locs:
-            lines.append(f"--- {loc.name} ---")
-            cached = weather_client.get_cached_weather(loc) if weather_client else None
-
-            if cached and cached.has_any_data() and cached.current:
-                cc = cached.current
-                settings = self.app.config_manager.get_settings()
-                temp_unit_pref = settings.temperature_unit
-                temp_unit = resolve_temperature_unit_preference(temp_unit_pref, loc)
-                round_values = getattr(settings, "round_values", False) if settings else False
-                precision = 0 if round_values else get_temperature_precision(temp_unit)
-                temp_str = format_temperature(
-                    cc.temperature_f,
-                    unit=temp_unit,
-                    temperature_c=cc.temperature_c,
-                    precision=precision,
-                )
-                cond_str = cc.condition or "Unknown"
-                lines.append(f"  Temperature: {temp_str}")
-                lines.append(f"  Condition: {cond_str}")
-
-                # Collect active alerts for this location.
-                active_alerts: list[object] = []
-                if cached.alerts is not None:
-                    if hasattr(cached.alerts, "get_active_alerts"):
-                        active_alerts = cached.alerts.get_active_alerts() or []
-                    elif hasattr(cached.alerts, "alerts"):
-                        active_alerts = cached.alerts.alerts or []
-
-                if active_alerts:
-                    lines.append(f"  Active Alerts: {len(active_alerts)}")
-                    for alert in active_alerts:
-                        event = getattr(alert, "event", "Unknown")
-                        severity = getattr(alert, "severity", "Unknown")
-                        lines.append(f"    • {event} ({severity})")
-                        location_alerts.append((loc.name, alert))
-                else:
-                    lines.append("  Active Alerts: None")
-
-                if getattr(cached, "stale", False):
-                    lines.append("  (Cached — data may be outdated)")
-            else:
-                lines.append("  (No cached data — select this location to load current conditions)")
-
-            lines.append("")
-
-        summary_text = "\n".join(lines)
-        self.current_conditions.SetValue(summary_text)
-
-        self._set_forecast_sections_visible(False)
-
-        self._all_locations_alerts_data = location_alerts
-        self._update_all_locations_alerts(location_alerts)
-
-        # Update tray icon with data from the highest-priority location.
-        tray_data, tray_loc_name = self._get_all_locations_tray_data()
-        self.app.update_tray_tooltip(tray_data, tray_loc_name)
-
-        self.stale_warning_label.SetLabel("")
-        MainWindow._update_precipitation_timeline_menu_state(self)
-
-        # Ensure the refresh button stays enabled in this view.
-        self.app.is_updating = False
-        self.refresh_button.Enable()
-
-    def _update_all_locations_alerts(self, location_alerts: list[tuple[str, object]]) -> None:
-        """
-        Populate the alerts list from aggregated all-locations alert data.
-
-        Each item is prefixed with the location name so screen-reader users
-        know which location triggered the alert without having to open details.
-
-        Args:
-            location_alerts: List of (location_name, WeatherAlert) pairs.
-
-        """
-        self.alerts_list.Clear()
-
-        if location_alerts:
-            items = []
-            for loc_name, alert in location_alerts:
-                event = getattr(alert, "event", "Unknown")
-                severity = getattr(alert, "severity", "Unknown")
-                items.append(f"{loc_name}: {event} ({severity})")
-            self.alerts_list.Append(items)
-            self.view_alert_button.Enable()
-        else:
-            self.view_alert_button.Disable()
-
-    def _get_all_locations_tray_data(self):
-        """
-        Return (weather_data, location_name) for the highest-priority location for tray display.
-
-        Priority order:
-        1. Location with the most severe active alert (Extreme > Severe > Moderate > Minor > Unknown)
-        2. First location that has any cached data (fallback when no alerts exist)
-
-        Returns None, None when no cached data is available at all.
-        """
-        SEVERITY_ORDER = ["Extreme", "Severe", "Moderate", "Minor", "Unknown"]
-
-        weather_client = (
-            self.app.weather_client
-            if hasattr(self.app, "weather_client") and self.app.weather_client
-            else None
-        )
-
-        try:
-            all_locs = [
-                loc
-                for loc in self.app.config_manager.get_all_locations()
-                if loc.name != "Nationwide"
-            ]
-        except Exception:
-            return None, None
-
-        best_data = None
-        best_name = None
-        best_severity_idx = len(SEVERITY_ORDER)  # sentinel: no alert yet
-
-        first_with_data = None
-        first_with_data_name = None
-
-        for loc in all_locs:
-            cached = weather_client.get_cached_weather(loc) if weather_client else None
-            if not cached or not cached.has_any_data():
-                continue
-
-            if first_with_data is None:
-                first_with_data = cached
-                first_with_data_name = loc.name
-
-            # Find the most severe alert for this location.
-            active_alerts: list = []
-            if cached.alerts is not None:
-                if hasattr(cached.alerts, "get_active_alerts"):
-                    active_alerts = cached.alerts.get_active_alerts() or []
-                elif hasattr(cached.alerts, "alerts"):
-                    active_alerts = cached.alerts.alerts or []
-
-            for alert in active_alerts:
-                severity = getattr(alert, "severity", None) or "Unknown"
-                if severity not in SEVERITY_ORDER:
-                    severity = "Unknown"
-                idx = SEVERITY_ORDER.index(severity)
-                if idx < best_severity_idx:
-                    best_severity_idx = idx
-                    best_data = cached
-                    best_name = loc.name
-
-        if best_data is not None:
-            return best_data, best_name
-
-        # Fallback: use the last single location the user was viewing
-        last_name = getattr(self, "_last_single_location_name", None)
-        if last_name and weather_client:
-            for loc in all_locs:
-                if loc.name == last_name:
-                    cached = weather_client.get_cached_weather(loc)
-                    if cached and cached.has_any_data():
-                        return cached, last_name
-
-        return first_with_data, first_with_data_name
-
-    def set_status(self, message: str) -> None:
-        """Set the status bar text and announce via screen reader."""
-        self.GetStatusBar().SetStatusText(message, 0)
-        logger.info(f"Status: {message}")
-        if message:
-            self._announcer.announce(message)
-
-    def _set_last_updated_status(self, when: datetime | None = None) -> None:
-        """
-        Write 'Last updated HH:MM' to the main status field silently.
-
-        Unlike set_status(), this does not invoke the screen-reader
-        announcer — refresh success is already conveyed by the
-        data_updated sound and a re-announcement on every refresh would
-        be noisy.  The timestamp is formatted in the user's local time
-        using 12-hour format to match Event Center entries.
-        """
-        stamp = (when or datetime.now()).strftime("%I:%M %p").lstrip("0")
-        self.GetStatusBar().SetStatusText(f"Last updated {stamp}", 0)
-
-    def _update_precipitation_timeline_menu_state(self, weather_data=None) -> None:
-        """Enable the precipitation timeline menu item when minutely data is available."""
-        menu_item = getattr(self, "_precipitation_timeline_item", None)
-        if menu_item is None:
-            return
-
-        is_enabled = False
-        if not getattr(self, "_all_locations_active", False):
-            data = (
-                weather_data
-                if weather_data is not None
-                else getattr(self.app, "current_weather_data", None)
-            )
-            forecast = getattr(data, "minutely_precipitation", None)
-            if forecast is not None:
-                has_data = getattr(forecast, "has_data", None)
-                if callable(has_data):
-                    is_enabled = bool(has_data())
-                else:
-                    is_enabled = bool(getattr(forecast, "points", None))
-
-        menu_item.Enable(is_enabled)
-
-    def _get_notification_event_manager(self):
-        """Get or create the notification event manager for AFD/severe risk notifications."""
-        return main_window_notification_events.get_notification_event_manager(self)
-
-    def _get_fallback_notifier(self):
-        """Get or create a cached fallback notifier for event notifications."""
-        return main_window_notification_events.get_fallback_notifier(self)
-
-    def _on_notification_event_data_received(self, weather_data) -> None:
-        """Handle lightweight event data without refreshing the visible weather UI."""
-        main_window_notification_events.on_notification_event_data_received(self, weather_data)
-
-    def _process_notification_events(self, weather_data) -> None:
-        """
-        Process weather data for notification events.
-
-        Checks for:
-        - Area Forecast Discussion (AFD) updates (NWS US only)
-        - Severe weather risk level changes
-
-        Both are opt-in notifications (disabled by default).
-        """
-        main_window_notification_events.process_notification_events(self, weather_data)
